refactor: nest GraphQL operations under a versionsCleaner namespace container
Move the 4 root Query + 5 root Mutation flat fields (versionsCleaner*) under one
hierarchical namespace container, per the Jahia GraphQL convention.

- Query:    versionsCleanerIsRunning -> versionsCleaner { isRunning }, etc.
- Mutation: versionsCleanerRun       -> versionsCleaner { run }, etc.

New holder types VersionsCleanerQuery/Mutation carry the operations as instance
@GraphQLField methods (the GqlVersionsCleanerConfig return type moves into the query
holder); the *Extension classes keep a single accessor field. @GraphQLRequiresPermission
stays on each leaf. Frontend gql.js + jsx and Cypress fixtures/specs updated in lockstep.
