+++ tests/cypress/fixtures/graphql/query/getConfig.graphql
@@ -1,11 +1,13 @@
 query {
-    versionsCleanerConfig {
-        disabled
-        cronExpression
-        nbVersionsToKeep
-        deleteOrphanedVersions
-        checkIntegrity
-        reindexDefaultWorkspace
-        maxExecutionTimeInMs
+    versionsCleaner {
+        config {
+            disabled
+            cronExpression
+            nbVersionsToKeep
+            deleteOrphanedVersions
+            checkIntegrity
+            reindexDefaultWorkspace
+            maxExecutionTimeInMs
+        }
     }
 }
