+++ tests/cypress/fixtures/graphql/query/getVersionCount.graphql
@@ -1,3 +1,5 @@
 query getVersionCount($nodePath: String) {
-    versionsCleanerVersionCount(nodePath: $nodePath)
+    versionsCleaner {
+        versionCount(nodePath: $nodePath)
+    }
 }
