+++ tests/cypress/fixtures/graphql/query/isRunning.graphql
@@ -1,3 +1,5 @@
 query {
-    versionsCleanerIsRunning
+    versionsCleaner {
+        isRunning
+    }
 }
